Profit changes

+++ app/Http/Requests/ProductRequest.php
@@ -31,7 +31,7 @@ class ProductRequest extends FormRequest
             'description' => ['required', 'string', 'between:3, 255'],
             'image' => ['nullable','file', 'mimes:jpg,jpeg,png','max:200000'],
             'base_cost' => ['numeric'],
-            'profit' => ['required','integer'],
+            'profit' => ['required','numeric'],
             'price' => ['numeric']
         ];
     }
